Add new 'killproc' OMZ plugin
This plugin provides the 'killproc()' function, which searches for input
process name for all processes belonging to that name, and sends
specific signals to them all; KILL initially, and if that doesn't work,
HUP and QUIT, etc.

The signals that are tried can be modified with the function parameters.
Also, processes can be limited to only those belonging to specific UID.